Move Status to libs
