+++ backend/index.js
@@ -6,7 +6,7 @@ const { spawn } = require('child_process');
 const { StatusCodes } = require('http-status-codes');
 const { wait, getEnabledCrawlerNames, getEnabledCrawlers } = require('./lib/utils');
 const config = require('./backend.config');
-const Status = require('./services/status');
+const Status = require('./lib/status');
 
 const app = express();
 const logger = pino();
